Fix bug on lines originating from another file
When a changed file contained lines that originated in another file (not
now being changed), that other file was treated as if it were also being
changed, and also its primary author was determined only by those lines
that got migrated into the changed file.

This change also eliminates the twisty Bash and Perl cruft in favor of
more intelligible Python code that forks git blame in subprocesses.